KPV-2583:: separate key extraction from cache impl
