+++ src/groovy/kuorum/core/navigation/cache/HasMapServletResponseCacheImpl.groovy
@@ -1,24 +1,11 @@
 package kuorum.core.navigation.cache
 
-import grails.plugin.springsecurity.SpringSecurityService
-import kuorum.core.customDomain.filter.ResponseContent
-import kuorum.register.KuorumUserSession
 
 import javax.servlet.ServletResponse
 
-class HasMapServletResponseCacheImpl implements ServletResponseCache {
-
-    SpringSecurityService springSecurityService
+class HasMapServletResponseCacheImpl extends AbstractUrlKeyCacheImpl implements ServletResponseCache {
 
     HashMap<String, ResponseContent> cache = new HashMap<>()
-
-    String buildKey(URL url) {
-        if (springSecurityService.isLoggedIn()) {
-            return url.toString() + ((KuorumUserSession) springSecurityService.principal).id
-        } else {
-            return url.toString();
-        }
-    }
 
     @Override
     void put(URL url, ResponseContent response) {
@@ -28,7 +15,7 @@ class HasMapServletResponseCacheImpl implements ServletResponseCache {
     @Override
     boolean get(URL url, ServletResponse response) {
         String key = buildKey(url)
-        if(cache.containsKey(key)){
+        if (cache.containsKey(key)) {
             cache.get(key).writeTo(response)
             return true
         }
